refactor: Function 모듈 이동
