--- src/main/java/codesquad/router/RouterFunction.java
@@ -1,10 +1,0 @@
-package codesquad.router;
-
-import codesquad.http11.HttpRequest;
-import codesquad.http11.HttpResponse;
-
-@FunctionalInterface
-public interface RouterFunction {
-
-    Object route(HttpRequest request, HttpResponse response);
-}
